spanner_cli/pyproject.toml
[project]
name = "scripts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.1.8",
    "google-cloud-spanner>=3.51.0",
    "pandas>=2.2.3",
    "pydantic>=2.10.5",
    "tqdm>=4.67.1",
]
